refactor(propose-task): convert to hook + dispatcher + Default archetype
Splits the ~765-line ProposeTask monolith into useProposeTask (the stable
ProposeTaskState contract: faction-list fetch, all form fields, length/metatask
guards, proposeTask vs proposeMetatask branch, success state) + a thin
dispatcher (keyed on the user-selected factionSlug via pickVariant, owning the
faction-agnostic login/eligibility gates) + DefaultProposeTask (the old form +
success screen lifted verbatim). No faction has a bespoke form yet, so every
faction renders the Default path unchanged.
